requirements.txt
beautifulsoup4==4.4.1
Django==1.9
enum34==1.1.2
gunicorn==19.4.5
idna==2.0
ipaddress==1.0.16
pyasn1==0.1.9
pycparser==2.14
python-dateutil==2.4.2
pytz==2015.7
six==1.10.0
wheel==0.29.0
whitenoise==2.0.6
